requirements.txt
# Core scientific stack (modern versions for Python 3.10)
numpy==1.26.4
scipy==1.14.1
scikit-learn==1.5.1
pandas==2.2.2
matplotlib==3.9.2

# Torch utils (compatible with Torch 2.4)
tensorboardX==2.6.2.2
torchsummaryX==1.3.0
torchinfo==1.8.0

# Config / plotting
pyaml==24.7.0
terminalplot==0.4

# Web / serving
flask==3.0.3
gunicorn==22.0.0

# ML libs
xgboost==2.1.1
lightgbm==4.5.0
